requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain_groq 
langchain_community 
faiss-cpu 
pymongo 
langchain-mongodb
huggingface_hub
sentence_transformers
pypdf 
docx2txt
python-multipart